+++ tools/generate_genesis.py
@@ -2,6 +2,7 @@ from __future__ import print_function
 
 import json
 
+import qrl.crypto.xmss
 from qrl.crypto import merkle
 
 num_accounts = 100
@@ -10,7 +11,7 @@ file_name = "aws_wallet"
 wallets = {}
 for i in range(num_accounts):
     print("Generating (", i + 1, "/", num_accounts, ")")
-    wallet = merkle.XMSS(signatures=4096, SEED=None)
+    wallet = qrl.crypto.xmss.XMSS(signatures=4096, SEED=None)
     wallets[wallet.address] = wallet.mnemonic
 
 with open(file_name, 'w') as f:
